refactor: 객체 분리 및 PR 피드백 반영
